Set the ubuntu 20.04 version, for use a more stable version.  Update the timezone to Sydney and add https certificates to resolve certificates issues  No need to use the 'add-apt-repository ppa:openjdk-r/ppa' as java is part of the ubuntu packages.  Improve script to clean up the /var/lib folder to reduce the image size
Set the ubuntu 20.04 version, for use a more stable version.

Update the timezone to Sydney and add https certificates to resolve certificates issues

No need to use the 'add-apt-repository ppa:openjdk-r/ppa' as java is part of the ubuntu packages.

Improve script to clean up the /var/lib folder to reduce the image size